modificado

+++ db/migrate/20241116184249_create_balances.rb
@@ -1,7 +1,7 @@
 class CreateBalances < ActiveRecord::Migration[7.2]
   def change
     create_table :balances do |t|
-      t.decimal :total_amount
+      t.float :total_amount
       t.boolean :status
       t.references :account, null: false, foreign_key: true
