+++ db/migrate/20241116185456_create_milestones.rb
@@ -1,7 +1,7 @@
 class CreateMilestones < ActiveRecord::Migration[7.2]
   def change
     create_table :milestones do |t|
-      t.decimal :amount
+      t.float :amount
       t.references :goal, null: false, foreign_key: true
 
       t.timestamps
