+++ db/migrate/20241129010841_add_amount_to_goal.rb
@@ -1,5 +1,5 @@
 class AddAmountToGoal < ActiveRecord::Migration[7.2]
   def change
-    add_column :goals, :amount, :decimal
+    add_column :goals, :amount, :float
   end
 end
